LayoutTests/imported/w3c/web-platform-tests/html/user-activation/activation-trigger-keyboard-enter.html

<!DOCTYPE html>
<html>
  <meta charset="utf-8" />
  <title>Test for keyboard activation trigger for ENTER key</title>
  <head>
    <link
      rel="help"
      href="https://html.spec.whatwg.org/multipage/interaction.html#activation-triggering-input-event"
    />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/resources/testdriver.js"></script>
    <script src="/resources/testdriver-vendor.js"></script>
    <script src="resources/utils.js"></script>
  </head>
  <body>
    <input type="text" autofocus />
  </body>
  <script>
    promise_test(async () => {
      const ENTER_KEY = "\uE007";

      const keydown_event = getEvent("keydown");
      const keypress_event = getEvent("keypress");
      const keyup_event = getEvent("keyup");

      await test_driver.send_keys(document.body, ENTER_KEY);

      await keydown_event;
      assert_true(
        navigator.userActivation.isActive,
        "ENTER keydown event should result in activation"
      );
      await consumeTransientActivation();

      await keypress_event;
      assert_false(
        navigator.userActivation.isActive,
        "ENTER keypress should have no activation after keydown consumption"
      );

      await keyup_event;
      assert_false(
        navigator.userActivation.isActive,
        "ENTER keyup should have no activation after keydown consumption"
      );
    }, "Activation through ENTER keyboard event");
  </script>
</html>
